+++ src/client/api/github.ts
@@ -1,6 +1,7 @@
 import { baseGHUrl } from "../constants";
 
-import { Job } from "../types";
+import { Job, LocationOption, SearchType } from "../types";
+import { unique } from "../util";
 
 export const getGitHubJobs = async (): Promise<Job[]> => {
   const response = await fetch(`${baseGHUrl}?location=los+angeles`, {
@@ -11,28 +12,42 @@ export const getGitHubJobs = async (): Promise<Job[]> => {
   return data;
 };
 
-export const searchJobs = async (searchString: string): Promise<Job[]> => {
-  const response = await fetch(
-    `${baseGHUrl}?description=${encodeURI(searchString)}`,
-    {
-      headers: { "Content-Type": "application/json" },
-      method: "GET",
-    }
-  );
+const getData = async (url: string): Promise<Job[]> => {
+  const response = await fetch(url, {
+    headers: { "Content-Type": "application/json" },
+    method: "GET",
+  });
   const data: Job[] = await response.json();
   return data;
 };
 
-export const locationSearchJobs = async (
-  locationSearchString: string
+export const searchJobs = async (
+  searchString: string,
+  locationOptions: LocationOption[],
+  searchType: SearchType
 ): Promise<Job[]> => {
-  const response = await fetch(
-    `${baseGHUrl}?location=${encodeURI(locationSearchString)}`,
-    {
-      headers: { "Content-Type": "application/json" },
-      method: "GET",
-    }
+  const jobs = [];
+
+  const locationsSearches = locationOptions.filter(
+    (location: LocationOption) => location.value !== ""
   );
-  const data: Job[] = await response.json();
-  return data;
+
+  await Promise.all(
+    locationsSearches.map(async (location: LocationOption) => {
+      let url = "";
+      if (searchType === "description") {
+        url = `${baseGHUrl}?description=${encodeURI(
+          searchString
+        )}&location=${encodeURI(location.value)}`;
+      } else if (searchType === "location") {
+        url = `${baseGHUrl}?location=${encodeURI(
+          searchString + " " + location.value
+        )}`;
+      }
+      const data = await getData(url);
+      jobs.push.apply(jobs, data);
+    })
+  );
+
+  return unique(jobs);
 };