+++ src/client/types.ts
@@ -13,3 +13,11 @@ export interface Job {
 }
 
 export type JobType = "Full Time";
+
+export interface LocationOption {
+  name: string;
+  setter: (param: string) => void;
+  value: string;
+}
+
+export type SearchType = "description" | "location";