+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-07-13
+
+### ✅ Select one option from at least 4 pre-defined options
+
+### Added
+
+- Location Options
+- `searchJobs()`
+- `util.ts`
+
+### Changed
+
+### Removed
+
+- Inidividual search function for location and description
+
+### Fixed
+
 ## [0.4.0] - 2020-07-13
 
 ### 🗺️ Search for jobs with a city name, zip code or other location